Organized the files structure by adding Auth folder inside the components folder.
